refactor(compare): split comparison sections
